Avoid using file append when installing code files
Some distributed file systems are not capable of opening files
in append mode. For instance, this happens on cloud systems,
on platforms like databricks, when trying to install a package
on a volume.

This leads to errors similar to

    Error writing to connection:  Operation not supported

when trying to install a package.

A current workaround is to install packages into a local
filesystem first, and then copy them to the distributed
filesystem.

I understand that non-POSIX compliant filesystems are not
officially supported, however I hope the patch can be merged
since it does not penalize existing users and makes package
installation more convenient in other environments.